+++ packages/angular-workspace/nimble-angular/table-column/duration-text/nimble-table-column-duration-text.directive.ts
@@ -1,7 +1,7 @@
 import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
 import { type TableColumnDurationText, tableColumnDurationTextTag } from '@ni/nimble-components/dist/esm/table-column/duration-text';
 import { BooleanValueOrAttribute, NumberValueOrAttribute, toBooleanProperty, toNullableNumberProperty } from '@ni/nimble-angular/internal-utilities';
-import { NimbleTableColumnBaseDirective } from '@ni/nimble-angular/table-column';
+import { NimbleTableColumnBaseDirective, TableColumnSortDirection } from '@ni/nimble-angular/table-column';
 
 export type { TableColumnDurationText };
 export { tableColumnDurationTextTag };
@@ -59,6 +59,30 @@ export class NimbleTableColumnDurationTextDirective extends NimbleTableColumnBas
 
     @Input('grouping-disabled') public set groupingDisabled(value: BooleanValueOrAttribute) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'groupingDisabled', toBooleanProperty(value));
+    }
+
+    public get sortingDisabled(): boolean {
+        return this.elementRef.nativeElement.sortingDisabled;
+    }
+
+    @Input('sorting-disabled') public set sortingDisabled(value: BooleanValueOrAttribute) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'sortingDisabled', toBooleanProperty(value));
+    }
+
+    public get sortDirection(): TableColumnSortDirection {
+        return this.elementRef.nativeElement.sortDirection;
+    }
+
+    @Input('sort-direction') public set sortDirection(value: TableColumnSortDirection) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'sortDirection', value);
+    }
+
+    public get sortIndex(): number | null | undefined {
+        return this.elementRef.nativeElement.sortIndex;
+    }
+
+    @Input('sort-index') public set sortIndex(value: NumberValueOrAttribute | null | undefined) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'sortIndex', toNullableNumberProperty(value));
     }
 
     public constructor(renderer: Renderer2, elementRef: ElementRef<TableColumnDurationText>) {