+++ packages/nimble-components/src/select/index.ts
@@ -40,7 +40,7 @@ import type { ErrorPattern } from '../patterns/error/types';
 import { iconExclamationMarkTag } from '../icons/exclamation-mark';
 import { isListOption, isListOptionGroup, template } from './template';
 import type { ListOption } from '../list-option';
-import { FilterMode } from './types';
+import { FilterMode, SelectFilterInputEventDetail } from './types';
 import { diacriticInsensitiveStringNormalizer } from '../utilities/models/string-normalizers';
 import { FormAssociatedSelect } from './models/select-form-associated';
 import type { ListOptionGroup } from '../list-option-group';
@@ -102,6 +102,9 @@ export class Select
 
     @attr({ attribute: 'clearable', mode: 'boolean' })
     public clearable = false;
+
+    @attr({ attribute: 'loading-visible', mode: 'boolean' })
+    public loadingVisible = false;
 
     /**
      * @internal
@@ -469,8 +472,8 @@ export class Select
     /**
      * @internal
      */
-    public inputClickHandler(e: MouseEvent): void {
-        e.stopPropagation(); // clicking in filter input shouldn't close dropdown
+    public ignoreClickHandler(e: MouseEvent): void {
+        e.stopPropagation();
     }
 
     /**
@@ -508,22 +511,30 @@ export class Select
      */
     public inputHandler(e: InputEvent): boolean {
         this.filter = this.filterInput?.value ?? '';
-        this.filterOptions();
+        if (this.filterMode === FilterMode.standard) {
+            this.filterOptions();
 
-        const enabledOptions = this.filteredOptions.filter(o => !o.disabled);
-        let activeOptionIndex = this.filter !== ''
-            ? this.openActiveIndex ?? this.selectedIndex
-            : this.selectedIndex;
+            const enabledOptions = this.filteredOptions.filter(
+                o => !o.disabled
+            );
+            let activeOptionIndex = this.filter !== ''
+                ? this.openActiveIndex ?? this.selectedIndex
+                : this.selectedIndex;
 
-        if (
-            enabledOptions.length > 0
-            && !enabledOptions.find(o => o === this.options[activeOptionIndex])
-        ) {
-            activeOptionIndex = this.options.indexOf(enabledOptions[0]!);
-        } else if (enabledOptions.length === 0) {
-            activeOptionIndex = -1;
+            if (
+                enabledOptions.length > 0
+                && !enabledOptions.find(o => o === this.options[activeOptionIndex])
+            ) {
+                activeOptionIndex = this.options.indexOf(enabledOptions[0]!);
+            } else if (enabledOptions.length === 0) {
+                activeOptionIndex = -1;
+            }
+            this.setActiveOption(activeOptionIndex);
         }
-        this.setActiveOption(activeOptionIndex);
+
+        if (this.filterMode !== FilterMode.none) {
+            this.emitFilterInputEvent();
+        }
 
         if (e.inputType.includes('deleteContent') || !this.filter.length) {
             return true;
@@ -854,6 +865,9 @@ export class Select
             this.filterInput.value = '';
         }
 
+        if (this.filterMode !== FilterMode.none) {
+            this.emitFilterInputEvent();
+        }
         this.ariaControls = '';
         this.ariaExpanded = 'false';
     }
@@ -1098,7 +1112,9 @@ export class Select
         if (option.hidden) {
             return true;
         }
-        return !this.filterMatchesText(option.text);
+        return this.filterMode === FilterMode.standard
+            ? !this.filterMatchesText(option.text)
+            : false;
     }
 
     private filterMatchesText(text: string): boolean {
@@ -1211,6 +1227,14 @@ export class Select
 
     private filterChanged(): void {
         this.filterOptions();
+    }
+
+    private emitFilterInputEvent(): void {
+        const eventDetail: SelectFilterInputEventDetail = {
+            filterText: this.filter
+        };
+
+        this.$emit('filter-input', eventDetail, { bubbles: true });
     }
 
     private maxHeightChanged(): void {